Core folder added with Interfaces and Repositories sub leves.
