+++ CreditCheck-BE/Data/IAuthRepository.cs
@@ -1,7 +1,4 @@
 ﻿using CreditCheck_BE.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace CreditCheck_BE.Data
